backend/requirements.txt
Django~=3.2.10
Pillow==9.2.0
Djangorestframework~=3.13.1
djangorestframework-simplejwt==5.2.1
djoser==2.1.0
django-cors-headers==3.13.0